+++ src/components/BulkRenewalDialog/BulkRenewedLoansList.js
@@ -11,7 +11,7 @@ import {
   Icon,
   MultiColumnList,
 } from '@folio/stripes/components';
-import { effectiveCallNumber } from '@folio/stripes-util';
+import { effectiveCallNumber } from '@folio/stripes/util';
 
 const propTypes = {
   height: PropTypes.number,